Helm chart: Allow passing full image name
Currently the Helm chart allows passing the repository and tag
separately, and concatenates the two in the deployment image field. This
works great when using an image, but when you want to pin your image
using your digest instead, it's not clear how this will work.

This patch adds a new Helm value, `image.name`, that, when specified,
will populate the entire image field. This allows users that want to
manage their full image themselves to do so.

Resolves #3018.