Re-structure the rhai logging output (#2975)
The Rhai logging output was "standardized" about a year ago. Since that
time, significant efforts have been made within the rest of the router
to standardize on a slightly different format. This PR brings rhai log
output into line with the rest of the router.

It also addresses the requirement to make the "message" output
specifiable by the script author.

Note: The changes are source code compatible, but the format of the
output log line is changing. I don't think we make compatibility
commitments on this, but it's worth noting.

Fixes #2777 

<!-- start metadata -->

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [ ] Changes are compatible[^1]
- [x] Documentation[^2] completed
- [x] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [x] Unit Tests
    - [x] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`