Do some hackery so we don't need RhaiHeaderMap
I believe it's acceptable (from a rhai point of view) to have each module
define its own type HeaderMap. When they are combined into a single module,
as long as the type definition is the same (not even sure if this is checked)
then everything works fine.

The change passes all of the tests and the example tests.

Also: polish up some of the functions which should be "pure". It will
perhaps give us a performance benefit.