Add a context try_into_iter() and use it in the co-processor (#3054)
also:
 - Modify upsert to use FnOnce rather than Fn to minimize cloning

@lennyburdette This is an attempt at a more performant fix than your PR
#3049. I think the end result is the same, it should just be slightly
more performant by avoiding cloning and also makes the base `Context`
slightly more efficient.

Could you try a build from this branch against your test case and
confirm it's still doing what you expect?

also: I've thought a little more about this and even though this is
"better" than what we have on `dev` right now, I still think there is
problem. This change means that we aren't racing across an entire
context, but we are still racing across subgraph responses for each
context update and that means there is no way to sequence the updating,
which is a possible problem. It really only becomes and issue with this
approach if both subgraphs are performing updates agains the same key,
so that's why it's a bit better than current `dev`, but it's still a
potential problem.


Fixes #3058 

<!-- start metadata -->

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [x] Changes are compatible[^1]
- [x] Documentation[^2] completed
- [x] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`