require_authentication plugin (#3002)
While the authentication plugin validates queries with JWT, it does not reject unauthenticated requests, and leaves that to other layers. This allows co-processors to handle other authentication methods, and plugins at later layers to authorize the request or not. Typically, this was done in rhai.

This now adds an option to the Router's YAML configuration to reject unauthenticated requests. It can be used as follows:

```yaml
authorization:
	require_authentication: true
```

The plugin will check for the presence of the `apollo_authentication::JWT::claims` key in the request context as proof that the request is authenticated.