chore: Refine release tooling to match `RELEASE_CHECKLIST.md` approach
This removes some automation which existed in the first iteration of the
releasing xtask which depended on `octorust`.

The steps in the updated `RELEASE_CHECKLIST.md` within
https://github.com/apollographql/router/pull/3110 have a list of commands
which leverage the `gh` CLI to do effecitely the same thing.

The instructions I've used for the last many releases seem to have validated
that approach and I suspect we will move in that direction.  As an added
benefit of depending on `gh`, `octorust` brought in a substantial amount of
build-time to the `xtask` crate and a _lot_ of dependencies.

(Concretely, I've been always having `--dry-run` and `--current-branch` on
with the current instructions, so those options are no longer that useful
there anyhow.)