extend IntoGraphqlErrors to include QueryPlannerError::Introspection
If we get an Introspection error during SupergraphService::plan_query(),
then it ends up being a 500 error. This change modifies the handling of
errors to generate a valid GraphQL error for Introspection errors.

The result of this change is that:

CLIENT RESPONSE:
StatusCode:500
{"errors":[{"message":"value retrieval failed: introspection error: introspection error : Field \"__schema\" of type \"__Schema!\" must have a selection of subfields. Did you mean \"__schema { ... }\"?","extensions":{"code":"INTERNAL_SERVER_ERROR"}}]}

BECOMES

CLIENT RESPONSE:
StatusCode:400
{"errors":[{"message":"introspection error : Field \"__schema\" of type \"__Schema!\" must have a selection of subfields. Did you mean \"__schema { ... }\"?","extensions":{"code":"INTROSPECTION_ERROR"}}]} + 400 StatusCode

In effect, all introspection errors become 400 rather than 500 and the
error message is slightly modified. I'm not sure if the extension code
should be "INTROSPECTION_ERROR" or something else. That's easy to
change.

fixes: #3090