experiment to be replace gnomic 500 errors with informative graphql
Currently, particularly if code is executing within a plugin, there are
many places within the router where we don't/can't return helpful
information to a client.

The result of such errors is something like this curl extract:

```
< HTTP/1.1 500 Internal Server Error
< content-type: text/plain; charset=utf-8
< content-length: 26
< date: Sat, 20 May 2023 10:35:52 GMT
<
* Connection #0 to host 127.0.0.1 left intact
router service call failed
```

Well, something has gone wrong and it could be pretty fatal. If you dig
about in the logs (assuming you have access to them) you might be able
to figure out what has gone wrong.

This experiment, provides a framework that we can leverage to gradually
add support for more meaninful/informative errors to the router. Using
the new `ContextError` type, we can return an informative error that,
for example, could be used to transform the above into:

```
< HTTP/1.1 200 OK
< content-type: application/json
< content-length: 163
< date: Tue, 23 May 2023 14:09:00 GMT
<
* Connection #0 to host 127.0.0.1 left intact
{"errors":[{"message":"could not parse response body into JSON value (invalid number at line 2 column 2)","extensions":{"code":"EXTERNAL_DESERIALIZATION_ERROR"}}]}
```

This is a much more informative error, that gives us a lot more
information about what has gone wrong.

`ContextError` allows the user to specify:
 - status code
 - contextual error message
 - extension code
 - source error

This information is then processed in the axum factory to generate the
meaningful error above. If there is no ContextError to be processed, you
get the existing behaviour, which allows us to progressively improve
things.

What does everyone think?