Add a very expressive list of commands to `RELEASE_CHECKLIST.md` (#3110)
Overall, this commit adds a very verbose list of individually automated
steps which can be automated into approximateily 3 fully-automated
chunks.

For now, we're starting with testing the process with manual commands,
and
this outlines those commands. I've been using these for the release for
a
while and literally just copy and paste the commands about after the 4th
step. (And stopping in the middle to fix the CHANGELOG.md and solicit PR
approval from the team).

Fixes https://github.com/apollographql/router/issues/170