coprocessor: add support for deferred responses and fix body data type (#3104)
Modifies the implementation to process deferred results gradually rather
than blocking.
Changes the coprocessor router body payload to be a String, rather than
a JSON Value (Map, effectively)

In retrospect, it was a mistake to have made the Router coprocessor body
payload be a JSON Value. It should just be a String (Bytes). There are
reasons for this:
 - Users want to interact with raw bytes at the Router level.
- Defer responses are structured, but aren’t JSON values and are a
stream.
 - Supergraph stage is where the body is guaranteed to be valid JSON.

Since we should modify the payload of the router coprocessor calls to be
bytes rather than JSON, we must do that as either:
A coprocessor protocol revision
A coprocessor protocol breaking change

If we had chosen 1, then we would need to decide how to implement
support in the router for multiple protocol versions (which doesn’t
exist as yet). Probably the best solution is to provide a configuration
option as part of coprocessor configuration and default to either the
oldest or newest (to be argued about) version.

This change, simply updates the implementation without revving the
protocol. Users will need to modify their coprocessors after router
version 1.x (when we release the fix).

Fixes #3015

<!-- start metadata -->

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [ ] Changes are compatible[^1]
- [x] Documentation[^2] completed
- [x] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [x] Unit Tests
    - [x] Integration Tests
    - [x] Manual Tests

**Exceptions**

Changes are **not** compatible, but we have made the decision to fix
this incompatibly now before the feature is widely used.
I created a simple co-processor which I could use to validate that my
changes were working as expected.

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`

---------

Co-authored-by: Stephen Barlow <stephen@apollographql.com>
Co-authored-by: Jesse Rosenberger <git@jro.cc>