don't try to convert bytes back to JSON on RouterRequest (#3138)
Now that we expect bytes back from the coprocessor, just take the bytes
and don't try to convert back to JSON.

There was an error in the RouterRequest processing logic. I was still
trying to convert coprocessor response bytes back to JSON and that
worked but resulted in double escaped text which failed when the query
planner tried to make sense of it.