Set the global allocator in the library crate, not just the executable (#3157)
Fixes https://github.com/apollographql/router/issues/3126

In 1.19, Apollo Router [switched to use jemalloc as the global Rust allocator on
Linux](https://github.com/apollographql/router/blob/dev/CHANGELOG.md#improve-memory-fragmentation-and-resource-consumption-by-switching-to-jemalloc-as-the-memory-allocator-on-linux-pr-2882)
to reduce memory fragmentation. However this is only done in the
executable binary provided by the `apollo-router` crate, so that [custom
binaries](https://www.apollographql.com/docs/router/customizations/custom-binary)
using the crate as a library were not affected.

Instead, the `apollo-router` library crate now sets the global allocator
so that custom binaries also take advantage of this by default. If some
other choice is desired, the `global-allocator` Cargo feature flag can
be disabled in `Cargo.toml` with:

```toml
[dependencies]
apollo-router = {version = "[…]", default-features = false}
```

Library crates that depend on `apollo-router` (if any) should also do
this in order to leave the choice to the eventual executable. (Cargo
default features are only disabled if *all* dependents specify
`default-features = false`.)

<!-- start metadata -->

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [ ] Changes are compatible[^1]
- [x] Documentation[^2] completed
- [x] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [x] Manual Tests

**Exceptions**

* If a custom binary was already using `#[global_allocator]`, upgrading
to this Router version will cause a compilation error until they disable
the Cargo feature flag as described above.
* Because of how default Cargo feature work, as explained in
https://github.com/apollographql/router/issues/3126, the test added in
this PR is in a separate workspace. Re-compiling the whole dependency
graph a second time on every CI run would be costly, so CI does not run
this test. I deem manual testing enough to check that `cfg` is spelled
correctly to make `#[global_allocator]` conditional on the relevant
Cargo feature.

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`