chore(deps): `xtask/` dependency updates, again
This is the result of running `cargo update` in the `xtask/` directory (our
directory of tooling; not runtime components) to bring things more up to
date.

Follows-up: https://github.com/apollographql/router/pull/3149

Some of the changes in https://github.com/apollographql/router/pull/3152
seem to have undone some of the dependency updates, so this re-applies them.