generate selections for the filtered query
Query planner plugins can modify the query before sending it through
query planning. They might require different sets of fields on the same
entity, and the query plan might add its own list of fields (keys for
federated queries).
As an example, We could have a User entity with its email field used as
key, and want to remove private information from the query.

So we could have the query:

{
  topProducts {
    name
    reviews {
      author {
        email
        username
      }
    }
  }
}

So here we would filter the query as follows:

{
  topProducts {
    name
    reviews {
      author {
        username
      }
    }
  }
}

But since the email is used as key, it would appear in the JSON object
that accumulates response data.
To avoid sending unrequested data to the client, we have the response
formatting phase, that uses selections extracted from the query, to pick
only what is needed from the JSON object, and send it to the client.

Here we need to apply the selections of the filtered query, to make sure
the email is not returned to the client (the original query would let it
go through). But we also need to apply the selections from the original
query, to have the null propagation algorithm apply and return a
response in the shape that matches the client query.

There is probably a better way to do it than applying selections twice,
but here we are sure the behaviour will be correct, and so far the
formatting phase is fast enough, we can spend a bit more time there