fix tests
 - add the new otel attributes to the prometheus tests
 - rename the match to include "_total" in the name

I'm not sure if that second change is correct. I think it is, because I
think the original behaviour (which appears to have changed) was a bit
looser and accepted a less specific name (looks like a synonym).

However, I'm not a prometheus file format expert, so I'll need to
verify that.