cache the compiler
creating the compiler from the query is an expensive operation, so its
result should be cached. The compiler cannot be cached in Redis, so we
only use the in memory cache, and the same number of entries as the
query planner cache.

There is a small architecture issue here: setting the `subselections` field and
executing the operation checks need mutable access to the Query object,
but if we cache it, we need to wrap it with an Arc, because its content
is not clonable, and the compiler itself is in a OnceCell<Mutex>.

Due to this, we actually need to regenerate the Query object entirely in
the bridge (the subselections can only be calculated once we have the
query plan)