Restore HTTP payload size limit, make it configurable (#3130)
Fixes https://github.com/apollographql/router/issues/2000

Configuration:

```yaml
preview_operation_limits:
  experimental_http_max_request_bytes: 2000000 # Default value: 2 MB
```

<!-- start metadata -->

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [ ] Changes are compatible[^1]
- [x] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [x] Integration Tests
    - [ ] Manual Tests

**Exceptions**

This change could be a compatibility issue for users who rely on large
requests and upgrading from a Router version that didn’t have this
limit. However this issue is easily fixed through configuration. The
error log (but not the HTTP response) mention the relevant configuration
key.

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`

---------

Co-authored-by: Chandrika Srinivasan <chandrikas@users.noreply.github.com>