add support for empty auth prefixes (#3206)
*Description here*

Fixes #2909. 

This fixes the `authentication.jwt` plugin to support empty prefixes for
the JWT header. Some companies use prefix-less headers; previously, the
authentication plugin would reject requests even with an empty header
explicitly set, such as:

```yml 
authentication:
  jwt:
    header_value_prefix: ""
```

This PR addresses that.

<!-- start metadata -->

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [x] Changes are compatible[^1]
- [] Documentation[^2] completed
- [] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [x] Unit Tests
    - [] Integration Tests
    - [x] Manual Tests

**Exceptions**

None.

**Notes**

Only note that the current iteration checks for whitespace to show the
`Invalid Prefix` error as to not show the base64 error message (and
avoid passing a raw value to the parser).

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`

---------

Co-authored-by: Bryn Cooke <BrynCooke@gmail.com>