add expansion module support to rhai
fixes: #1744

This refactors existing functionality within `Expansion` to add support
for a default rhai expander and to expose a new function for environment
variable access. This leverages the existing mechanism for limiting
access to environment variables.

This functionality is then exposed to rhai scripts as in this example
code:

```
    let expander = expansion::create();
    try {
        print(`LANG: ${expansion::env(expander, "LANG")}`);
    } catch(err) {
        print(`exception: ${err}`);
    }
```

In the sample, we use the new expansion module to create an expander. We
then use the expander to access environment variables, in this case
"LANG".